Enemyの動作をタイムスケールに対応

+++ source/System/EnemyTween.h
@@ -42,13 +42,13 @@ public:
 	引数 type : イージングタイプ
 	引数 callback : 終了時のコールバック関数
 	***************************************/
-	static void Move(BaseEnemy& ref, const D3DXVECTOR3& startPosition, const D3DXVECTOR3& endPosition, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Move(BaseEnemy& ref, const D3DXVECTOR3& startPosition, const D3DXVECTOR3& endPosition, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	移動トゥイーン
 	基本的には上記と一緒だが、移動開始座標を現在座標に自動で設定してくれる
 	***************************************/
-	static void Move(BaseEnemy& ref, const D3DXVECTOR3& endPosition, int duratino, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Move(BaseEnemy& ref, const D3DXVECTOR3& endPosition, float duratino, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	スケールトゥイーン
@@ -59,13 +59,13 @@ public:
 	引数 type : イージングタイプ
 	引数 callback : 終了時のコールバック関数
 	***************************************/
-	static void Scale(BaseEnemy& ref, const D3DXVECTOR3& startScale, const D3DXVECTOR3& endScale, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Scale(BaseEnemy& ref, const D3DXVECTOR3& startScale, const D3DXVECTOR3& endScale, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	スケールトゥイーン
 	基本的には上記と一緒だが開始時のスケールを現在スケールに自動で設定してくれる
 	***************************************/
-	static void Scale(BaseEnemy& ref, const D3DXVECTOR3& endScale, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Scale(BaseEnemy& ref, const D3DXVECTOR3& endScale, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	回転トゥイーン
@@ -77,13 +77,13 @@ public:
 	引数 callback : 終了時のコールバック関数
 	※回転角度は自動で0.0f~360.0fに補正されてしまうので注意
 	***************************************/
-	static void Rotate(BaseEnemy& ref, const D3DXVECTOR3& startEulerAngle, const D3DXVECTOR3& endEulerAngle, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Rotate(BaseEnemy& ref, const D3DXVECTOR3& startEulerAngle, const D3DXVECTOR3& endEulerAngle, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	回転トゥイーン
 	基本的には上記と一緒だが開始の回転角度を自動で現在角度に設定してくれる
 	***************************************/
-	static void Rotate(BaseEnemy& ref, const D3DXVECTOR3& endEulaerAngle, int duration, EaseType type, std::function<void(void)> callback = nullptr);
+	static void Rotate(BaseEnemy& ref, const D3DXVECTOR3& endEulaerAngle, float duration, EaseType type, std::function<void(void)> callback = nullptr);
 
 	/**************************************
 	方向トゥイーン
@@ -94,7 +94,7 @@ public:
 	引数 dummyAxis : 始点の向きと終点の向きが平行だった場合に使用する回転軸
 	引数 callback : 終了時のコールバック関数
 	***************************************/
-	static void Turn(BaseEnemy& ref, const D3DXVECTOR3& endDirection, int duration, EaseType type, const D3DXVECTOR3& dummyAxis, std::function<void()> callback = nullptr);
+	static void Turn(BaseEnemy& ref, const D3DXVECTOR3& endDirection, float duration, EaseType type, const D3DXVECTOR3& dummyAxis, std::function<void()> callback = nullptr);
 
 private:
 	void Update();
@@ -113,7 +113,7 @@ private:
 	class EnemyTweener
 	{
 	public:
-		EnemyTweener(std::shared_ptr<Transform>& ref, int duration, EaseType type, Callback callback);
+		EnemyTweener(std::shared_ptr<Transform>& ref, float duration, EaseType type, Callback callback);
 		virtual ~EnemyTweener();
 		inline bool IsFinished();
 		virtual void Update() = 0;
@@ -121,8 +121,8 @@ private:
 
 	protected:
 		std::weak_ptr<Transform> reference;
-		int cntFrame;
-		int duration;
+		float cntFrame;
+		float duration;
 		EaseType type;
 		Callback callback;
 	};
@@ -130,7 +130,7 @@ private:
 	class MoveEnemyTweener : public EnemyTweener
 	{
 	public:
-		MoveEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXVECTOR3& start, const D3DXVECTOR3& end, int duration, EaseType type, Callback callback);
+		MoveEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXVECTOR3& start, const D3DXVECTOR3& end, float duration, EaseType type, Callback callback);
 		void Update();
 
 	private:
@@ -140,7 +140,7 @@ private:
 	class ScaleEnemyTweener : public EnemyTweener
 	{
 	public:
-		ScaleEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXVECTOR3& start, const D3DXVECTOR3& end, int duration, EaseType type, Callback callback);
+		ScaleEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXVECTOR3& start, const D3DXVECTOR3& end, float duration, EaseType type, Callback callback);
 		void Update();
 
 	private:
@@ -150,8 +150,8 @@ private:
 	class RotateEnemyTweener : public EnemyTweener
 	{
 	public:
-		RotateEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXVECTOR3& start, const D3DXVECTOR3& end, int duration, EaseType type, Callback callback);
-		RotateEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXQUATERNION& start, const D3DXQUATERNION& end, int duration, EaseType type, Callback callback);
+		RotateEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXVECTOR3& start, const D3DXVECTOR3& end, float duration, EaseType type, Callback callback);
+		RotateEnemyTweener(std::shared_ptr<Transform>& ref, const D3DXQUATERNION& start, const D3DXQUATERNION& end, float duration, EaseType type, Callback callback);
 		void Update();
 
 	private:
